PageBaseをパッケージBaseに追加
